cli: rename `queue` → `message-queue` + add `delete` leaf (#211)
Rename the entire queue namespace to message-queue across the SDK
and CLI surfaces. The current name was too generic — a future
queue of anything else (functions, swarms, tools) would collide.
What's queued is always a message-shaped RichContent, so
"message-queue" makes intent obvious in --help and JSON-Schema
type names.

Naming conventions (split by surface, matching agents/read/id):
  CLI argv:                "message-queue"
  Rust module path:        "message_queue"
  serde Path rename:       "message-queue"
  schemars rename:         "message_queue"

Affected surfaces:
- SDK: agents::queue → agents::message_queue (11 files, plus
  Queue → MessageQueue variants on Command / Request /
  ResponseItem at agents/mod.rs).
- CLI: command::agents::queue → command::agents::message_queue
  (5 files, plus dispatch rewire at agents/mod.rs).
- CLI: command::queue_drain → command::message_queue_drain
  (helpers join_with_separator + combine_drain_failure stay
  named for what they do, not which queue they serve).
- Storage: db::prompts module and the `prompts` table stay
  named — internal-only, no migration story needed.
- One stray cross-tier ref in filesystem/logs/client.rs
  (queue_list helper return type) updated.

New leaf: `agents message-queue delete <id>`
- Atomically drops one queued prompt by its prompts.id (as
  surfaced by `agents message-queue list`).
- Returns { id, agent_instance_hierarchy, agent_tag,
  enqueued_at, content: RichContent } so callers can confirm
  the dropped item.
- Storage helper delete_by_id_async reuses the existing
  collect_matching_prompts + reconstruct_and_delete pair with
  a single-row WHERE p.id = ? predicate. Cascade on
  prompt_contents.prompt_id sweeps content rows in the same
  transaction.
- Wired through the same Command/Request/ResponseItem dispatch
  tree as add/list/read; request_schema + response_schema
  siblings mirror the established pattern.

Doc-comment churn: leading module headers and shared-helper
references swapped from "agents queue ..." to
"agents message-queue ..." where natural; the colloquial "queue"
stays in surrounding prose since it still names the concept
correctly.
