+++ objectiveai-cli/src/command/agents/spawn.rs
@@ -1,19 +1,19 @@
-//! `agents spawn` — bare-naked chunk-or-id streaming handler.
+//! `agents spawn` â€” bare-naked chunk-or-id streaming handler.
 //!
 //! Spawns the instance runner as a detached background process. The
 //! streamed item shape depends on `dangerous_advanced.stream`:
 //!
-//! - **`None | Some(false)` (default)** — yields exactly one
+//! - **`None | Some(false)` (default)** â€” yields exactly one
 //!   `ResponseItem::Id(<agent_instance_hierarchy>)` once the runner's
 //!   `LogStreamReady` handshake fires, then the leaf's stream ends.
 //!   The instance runner child keeps running orphaned and drives the
 //!   completion to completion.
-//! - **`Some(true)`** — yields the same `ResponseItem::Id` first, then
+//! - **`Some(true)`** â€” yields the same `ResponseItem::Id` first, then
 //!   one `ResponseItem::Chunk` per chunk Notification the runner
 //!   emits, until the runner's stdout EOFs.
 //!
 //! `params.stream` on the wire is always `Some(true)` regardless of the
-//! `dangerous_advanced.stream` field — the latter controls the leaf's
+//! `dangerous_advanced.stream` field â€” the latter controls the leaf's
 //! output stream behaviour, not the API request.
 
 use std::path::PathBuf;
@@ -44,7 +44,7 @@ pub async fn execute(ctx: &Context, request: Request) -> Result<ItemStream, Erro
     // predicate (see `db::prompts::drain_for_spawn_async`):
     //   1. queue items addressed to `request.agent_tag` directly
     //   2. queue items addressed to any PENDING tag whose
-    //      (parent, agent_full_id) matches this spawn — i.e. tags
+    //      (parent, agent_full_id) matches this spawn â€” i.e. tags
     //      that the first-chunk auto-promotion hook will bind.
     // The drained items get joined oldest-first with `\n\n`
     // separators and prepended as a fresh User message at the head
@@ -58,7 +58,7 @@ pub async fn execute(ctx: &Context, request: Request) -> Result<ItemStream, Erro
     )
     .await?;
     if !drained.is_empty() {
-        let prepended = crate::command::queue_drain::join_with_separator(
+        let prepended = crate::command::message_queue_drain::join_with_separator(
             drained.iter().map(|d| d.content.clone()).collect(),
         );
         messages.insert(
@@ -100,7 +100,7 @@ pub async fn execute(ctx: &Context, request: Request) -> Result<ItemStream, Erro
     // anything), restore the drained queue rows via
     // `db::prompts::re_enqueue_async` and surface the original
     // error. On Ok, hand back a `StreamOnce::new(head).chain(tail)`
-    // — same pattern objectiveai-api's
+    // â€” same pattern objectiveai-api's
     // `functions/executions/client.rs::create_streaming_handle_usage`
     // uses for peek-then-stream returns.
     let mut tail: ItemStream = Box::pin(raw.map(map_item));
@@ -110,11 +110,11 @@ pub async fn execute(ctx: &Context, request: Request) -> Result<ItemStream, Erro
         )),
         Some(Err(e)) => {
             let r = db::prompts::re_enqueue_async(ctx.filesystem.clone(), drained).await;
-            Err(crate::command::queue_drain::combine_drain_failure(e, r))
+            Err(crate::command::message_queue_drain::combine_drain_failure(e, r))
         }
         None => {
             let r = db::prompts::re_enqueue_async(ctx.filesystem.clone(), drained).await;
-            Err(crate::command::queue_drain::combine_drain_failure(
+            Err(crate::command::message_queue_drain::combine_drain_failure(
                 Error::EmptyStream,
                 r,
             ))
@@ -178,7 +178,7 @@ async fn resolve_agent(
 /// PENDING tags that this spawn will bind on first chunk.
 ///
 /// - `AgentSpec::Resolved(InlineAgentBaseWithFallbacks)`: compute
-///   locally via `.convert().full_id()` — no HTTP.
+///   locally via `.convert().full_id()` â€” no HTTP.
 /// - `AgentSpec::Resolved(Remote)` and `AgentSpec::Favorite`: fetch
 ///   the remote definition via the same `objectiveai_sdk::agent::
 ///   get_agent` call `agents get` uses, then call `.full_id()` on
