+++ objectiveai-cli/src/filesystem/client.rs
@@ -8,17 +8,17 @@ pub struct Client {
     pub commit_author_email: String,
     /// Lazily-initialised SQLite connection shared across clones of this
     /// `Client`. Populated on first call to `filesystem::db::connection::connection`.
-    /// Wrapped in `Arc<Mutex<Option<…>>>` so every clone sees the same
+    /// Wrapped in `Arc<Mutex<Option<â€¦>>>` so every clone sees the same
     /// connection once one exists, and so init failures don't poison
-    /// the slot — a failed attempt leaves the inner `Option::None`
+    /// the slot â€” a failed attempt leaves the inner `Option::None`
     /// intact and later calls can retry.
     db_conn: Arc<Mutex<Option<Arc<Mutex<rusqlite::Connection>>>>>,
-    /// Parallel slot for the dedicated `tags.sqlite` connection — same
+    /// Parallel slot for the dedicated `tags.sqlite` connection â€” same
     /// lazy-init semantics as `db_conn` but a separate file so the
     /// agent-tags lifecycle stays isolated from the main message-log
     /// database. Now also hosts the `prompts` table for
-    /// `agents queue {add,list}` so the queue-list leaf can JOIN
-    /// prompts ⨝ tags in one query.
+    /// `agents message-queue {add,list}` so the queue-list leaf can JOIN
+    /// prompts â¨ tags in one query.
     tags_db_conn: Arc<Mutex<Option<Arc<Mutex<rusqlite::Connection>>>>>,
 }
 
@@ -72,7 +72,7 @@ impl Client {
     }
 
     /// Path to the dedicated `tags.sqlite` file under `base_dir`.
-    /// Also holds the `prompts` table for `agents queue {add,list}`.
+    /// Also holds the `prompts` table for `agents message-queue {add,list}`.
     pub fn tags_db_path(&self) -> PathBuf {
         self.base_dir.join("tags.sqlite")
     }
