+++ objectiveai-cli/src/filesystem/db/tags.rs
@@ -7,9 +7,9 @@
 //! One table, `tags`. Each row is in exactly one of two states (a
 //! `CHECK` constraint enforces mutual exclusion):
 //!
-//! - **BOUND** — `agent_instance_hierarchy` set, the other two NULL.
+//! - **BOUND** â€” `agent_instance_hierarchy` set, the other two NULL.
 //!   The tag points at one specific hierarchy.
-//! - **PENDING** — `parent_agent_instance_hierarchy + agent_full_id`
+//! - **PENDING** â€” `parent_agent_instance_hierarchy + agent_full_id`
 //!   set, `agent_instance_hierarchy` NULL. The tag is waiting for the
 //!   next agent-completion that matches the (full_id, parent) pair to
 //!   spawn under it, at which point its first chunk auto-promotes the
@@ -53,8 +53,8 @@ pub fn connection(client: &Client) -> Result<Arc<Mutex<Connection>>, Error> {
 
 /// Create every table this DB hosts. `tags.sqlite` started as a
 /// single-table file but now also holds `prompts` (the deferred-
-/// message queue from `agents queue add`) so the queue-list leaf
-/// can JOIN prompts ⨝ tags in a single SELECT.
+/// message queue from `agents message-queue add`) so the queue-list leaf
+/// can JOIN prompts â¨ tags in a single SELECT.
 fn init_tables(conn: &Connection) -> Result<(), Error> {
     conn.execute_batch(
         "CREATE TABLE IF NOT EXISTS tags (\
@@ -195,7 +195,7 @@ pub fn upsert_bound(
 
 /// All tags currently bound to the given hierarchy, newest-bound
 /// first. A hierarchy can carry multiple tag names (the schema's PK
-/// is the `name`, not the hierarchy) — every BOUND row whose
+/// is the `name`, not the hierarchy) â€” every BOUND row whose
 /// `agent_instance_hierarchy` matches is returned. PENDING rows
 /// never match.
 pub fn tags_for_hierarchy(
@@ -299,7 +299,7 @@ pub fn lookup(client: &Client, name: &str) -> Result<LookupState, Error> {
 
 /// Parent scope of an `agent_instance_hierarchy`: the substring up
 /// to (but not including) the last `/`. When the input has no `/`,
-/// the parent is the empty string — the hierarchy is at the root.
+/// the parent is the empty string â€” the hierarchy is at the root.
 pub(crate) fn parent_of(agent_instance_hierarchy: &str) -> &str {
     match agent_instance_hierarchy.rfind('/') {
         Some(i) => &agent_instance_hierarchy[..i],
@@ -487,7 +487,7 @@ mod tests {
         let (c, _tmp) = fresh_client();
         // Pending registered against parent `root/A`.
         upsert_pending(&c, "foo", "F", "root/A").unwrap();
-        // Spawn fires with hierarchy `root/A/inst-1` — parent is `root/A`.
+        // Spawn fires with hierarchy `root/A/inst-1` â€” parent is `root/A`.
         let promoted = upgrade(&c, "F", "root/A/inst-1").unwrap();
         assert_eq!(promoted, vec!["foo".to_string()]);
         // Same agent_full_id under a different parent should NOT promote
@@ -501,7 +501,7 @@ mod tests {
     fn upgrade_at_root_uses_empty_parent() {
         let (c, _tmp) = fresh_client();
         upsert_pending(&c, "rootless", "F", "").unwrap();
-        // Hierarchy with no `/` → parent = "".
+        // Hierarchy with no `/` â†’ parent = "".
         let promoted = upgrade(&c, "F", "lonely-root").unwrap();
         assert_eq!(promoted, vec!["rootless".to_string()]);
         assert_eq!(
