+++ objectiveai-sdk-rs/src/cli/command/agents/message_queue/delete/mod.rs
@@ -1,0 +1,131 @@
+//! `agents message-queue delete` — manually drop one queued prompt
+//! by its `prompts.id` (the id surfaced by `agents message-queue
+//! list`). Returns the deleted row's metadata + content so callers
+//! can confirm exactly which item was dropped.
+//!
+//! Cascade on `prompt_contents.prompt_id` sweeps every per-kind
+//! content row inside the same transaction, so the delete is
+//! atomic.
+
+use crate::cli::command::CommandRequest;
+
+#[derive(Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize, schemars::JsonSchema)]
+#[schemars(rename = "cli.command.agents.message_queue.delete.Request")]
+pub struct Request {
+    pub path_type: Path,
+    pub id: i64,
+    pub jq: Option<String>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize, schemars::JsonSchema)]
+#[schemars(rename = "cli.command.agents.message_queue.delete.Path")]
+pub enum Path {
+    #[serde(rename = "agents/message-queue/delete")]
+    AgentsMessageQueueDelete,
+}
+
+impl CommandRequest for Request {
+    fn into_command(&self) -> Vec<String> {
+        let mut argv = vec![
+            "agents".to_string(),
+            "message-queue".to_string(),
+            "delete".to_string(),
+            self.id.to_string(),
+        ];
+        if let Some(jq) = &self.jq {
+            argv.push("--jq".to_string());
+            argv.push(jq.clone());
+        }
+        argv
+    }
+}
+
+/// What was deleted. Carries every column of the original
+/// `prompts` row so the caller can confirm the drop:
+/// exactly one of `agent_instance_hierarchy` / `agent_tag` is set
+/// (matching the original target), `enqueued_at` is the original
+/// unix-seconds timestamp, and `content` is the reconstructed
+/// `RichContent` body.
+#[derive(Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize, schemars::JsonSchema)]
+#[schemars(rename = "cli.command.agents.message_queue.delete.Response")]
+pub struct Response {
+    pub id: i64,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    #[schemars(extend("omitempty" = true))]
+    pub agent_instance_hierarchy: Option<String>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    #[schemars(extend("omitempty" = true))]
+    pub agent_tag: Option<String>,
+    pub enqueued_at: i64,
+    pub content: crate::agent::completions::message::RichContent,
+}
+
+#[derive(clap::Args)]
+pub struct Args {
+    /// Row id of the queued prompt to delete (as surfaced by
+    /// `agents message-queue list`).
+    pub id: i64,
+    /// jq filter applied to the JSON output.
+    #[arg(long)]
+    pub jq: Option<String>,
+}
+
+#[derive(clap::Args)]
+#[command(args_conflicts_with_subcommands = true)]
+pub struct Command {
+    #[command(flatten)]
+    pub args: Args,
+    #[command(subcommand)]
+    pub schema: Option<Schema>,
+}
+
+#[derive(clap::Subcommand)]
+pub enum Schema {
+    /// Emit the JSON Schema for this leaf's `Request` type and exit.
+    RequestSchema(request_schema::Args),
+    /// Emit the JSON Schema for this leaf's `Response` type and exit.
+    ResponseSchema(response_schema::Args),
+}
+
+impl TryFrom<Args> for Request {
+    type Error = crate::cli::command::FromArgsError;
+    fn try_from(args: Args) -> Result<Self, Self::Error> {
+        Ok(Self {
+            path_type: Path::AgentsMessageQueueDelete,
+            id: args.id,
+            jq: args.jq,
+        })
+    }
+}
+
+#[cfg(feature = "cli-executor")]
+pub async fn execute<E: crate::cli::command::CommandExecutor>(
+    executor: &E,
+    mut request: Request,
+    agent_arguments: Option<&crate::cli::command::AgentArguments>,
+) -> Result<Response, E::Error> {
+    request.jq = None;
+    executor.execute_one(request, agent_arguments).await
+}
+
+#[cfg(feature = "cli-executor")]
+pub async fn execute_jq<E: crate::cli::command::CommandExecutor>(
+    executor: &E,
+    mut request: Request,
+    jq: String,
+    agent_arguments: Option<&crate::cli::command::AgentArguments>,
+) -> Result<serde_json::Value, E::Error> {
+    request.jq = Some(jq);
+    executor.execute_one(request, agent_arguments).await
+}
+
+#[cfg(feature = "mcp")]
+impl crate::cli::command::CommandResponse for Response {
+    fn into_mcp(self) -> crate::cli::command::McpResponseItem {
+        crate::cli::command::McpResponseItem::JSONL(serde_json::to_value(self).unwrap())
+    }
+}
+
+pub mod request_schema;
+
+pub mod response_schema;
