+++ objectiveai-sdk-rs/src/cli/command/agents/message_queue/read/mod.rs
@@ -1,0 +1,141 @@
+//! `agents message-queue read` sub-tier. One leaf today: `id`.
+
+use crate::cli::command::CommandRequest;
+
+pub mod id;
+
+#[derive(clap::Subcommand)]
+pub enum Command {
+    /// Fetch one piece of queued content by `prompt_contents.id`.
+    Id(id::Command),
+}
+
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, schemars::JsonSchema)]
+#[serde(untagged)]
+#[schemars(rename = "cli.command.agents.message_queue.read.Request")]
+pub enum Request {
+    #[schemars(title = "Id")]
+    Id(id::Request),
+    #[schemars(title = "IdRequestSchema")]
+    IdRequestSchema(id::request_schema::Request),
+    #[schemars(title = "IdResponseSchema")]
+    IdResponseSchema(id::response_schema::Request),
+}
+
+// Exempt from json-schema coverage: tier aggregate (see the root
+// `ResponseItem` in command.rs - TS7056).
+#[objectiveai_sdk_macros::json_schema_ignore]
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, schemars::JsonSchema)]
+#[schemars(rename = "cli.command.agents.message_queue.read.ResponseItem")]
+#[serde(untagged)]
+pub enum ResponseItem {
+    #[schemars(title = "Id")]
+    Id(id::Response),
+    #[schemars(title = "IdRequestSchema")]
+    IdRequestSchema(id::request_schema::Response),
+    #[schemars(title = "IdResponseSchema")]
+    IdResponseSchema(id::response_schema::Response),
+}
+
+#[cfg(feature = "mcp")]
+impl crate::cli::command::CommandResponse for ResponseItem {
+    fn into_mcp(self) -> crate::cli::command::McpResponseItem {
+        match self {
+            ResponseItem::Id(v) => v.into_mcp(),
+            ResponseItem::IdRequestSchema(v) => v.into_mcp(),
+            ResponseItem::IdResponseSchema(v) => v.into_mcp(),
+        }
+    }
+}
+
+impl TryFrom<Command> for Request {
+    type Error = crate::cli::command::FromArgsError;
+    fn try_from(command: Command) -> Result<Self, Self::Error> {
+        match command {
+            Command::Id(cmd) => match cmd.schema {
+                None => Ok(Request::Id(id::Request::try_from(cmd.args)?)),
+                Some(id::Schema::RequestSchema(args)) => Ok(
+                    Request::IdRequestSchema(id::request_schema::Request::try_from(args)?),
+                ),
+                Some(id::Schema::ResponseSchema(args)) => Ok(
+                    Request::IdResponseSchema(id::response_schema::Request::try_from(args)?),
+                ),
+            },
+        }
+    }
+}
+
+impl CommandRequest for Request {
+    fn into_command(&self) -> Vec<String> {
+        match self {
+            Request::Id(inner) => inner.into_command(),
+            Request::IdRequestSchema(inner) => inner.into_command(),
+            Request::IdResponseSchema(inner) => inner.into_command(),
+        }
+    }
+}
+
+#[cfg(feature = "cli-executor")]
+pub async fn execute<E: crate::cli::command::CommandExecutor>(
+    executor: &E,
+    request: Request,
+    agent_arguments: Option<&crate::cli::command::AgentArguments>,
+) -> Result<
+    std::pin::Pin<Box<dyn futures::Stream<Item = Result<ResponseItem, E::Error>> + Send>>,
+    E::Error,
+> {
+    let stream: std::pin::Pin<
+        Box<dyn futures::Stream<Item = Result<ResponseItem, E::Error>> + Send>,
+    > = match request {
+        Request::Id(req) => {
+            let value = id::execute(executor, req, agent_arguments).await?;
+            Box::pin(crate::cli::command::StreamOnce::new(Ok(ResponseItem::Id(value))))
+        }
+        Request::IdRequestSchema(req) => {
+            let value =
+                id::request_schema::execute(executor, req, agent_arguments).await?;
+            Box::pin(crate::cli::command::StreamOnce::new(Ok(
+                ResponseItem::IdRequestSchema(value),
+            )))
+        }
+        Request::IdResponseSchema(req) => {
+            let value =
+                id::response_schema::execute(executor, req, agent_arguments).await?;
+            Box::pin(crate::cli::command::StreamOnce::new(Ok(
+                ResponseItem::IdResponseSchema(value),
+            )))
+        }
+    };
+    Ok(stream)
+}
+
+#[cfg(feature = "cli-executor")]
+pub async fn execute_jq<E: crate::cli::command::CommandExecutor>(
+    executor: &E,
+    request: Request,
+    jq: String,
+    agent_arguments: Option<&crate::cli::command::AgentArguments>,
+) -> Result<
+    std::pin::Pin<Box<dyn futures::Stream<Item = Result<serde_json::Value, E::Error>> + Send>>,
+    E::Error,
+> {
+    let stream: std::pin::Pin<
+        Box<dyn futures::Stream<Item = Result<serde_json::Value, E::Error>> + Send>,
+    > = match request {
+        Request::Id(req) => {
+            let value = id::execute_jq(executor, req, jq, agent_arguments).await?;
+            Box::pin(crate::cli::command::StreamOnce::new(Ok(value)))
+        }
+        Request::IdRequestSchema(req) => {
+            let value =
+                id::request_schema::execute_jq(executor, req, jq, agent_arguments).await?;
+            Box::pin(crate::cli::command::StreamOnce::new(Ok(value)))
+        }
+        Request::IdResponseSchema(req) => {
+            let value =
+                id::response_schema::execute_jq(executor, req, jq, agent_arguments).await?;
+            Box::pin(crate::cli::command::StreamOnce::new(Ok(value)))
+        }
+    };
+    Ok(stream)
+}
