+++ objectiveai-sdk-rs/src/cli/command/agents/mod.rs
@@ -3,7 +3,7 @@ pub mod list;
 pub mod me;
 pub mod message;
 pub mod publish;
-pub mod queue;
+pub mod message_queue;
 pub mod read;
 pub mod spawn;
 pub mod tags;
@@ -12,7 +12,7 @@ pub mod tags;
 pub enum Command {
     /// Get an agent by remote path or favorite name.
     Get(get::Command),
-    /// List agents — `active` (direct children of the calling agent) or
+    /// List agents ÃƒÂ¢Ã¢â€šÂ¬Ã¢â‚¬Â `active` (direct children of the calling agent) or
     /// `available` (remote agents by source).
     List {
         #[command(subcommand)]
@@ -25,10 +25,10 @@ pub enum Command {
     Message(message::Command),
     /// Publish an agent to the local filesystem.
     Publish(publish::Command),
-    /// Deferred-prompt queue — `add` (more leaves to follow).
-    Queue {
+    /// Deferred-prompt queue ÃƒÂ¢Ã¢â€šÂ¬Ã¢â‚¬Â `add` (more leaves to follow).
+    MessageQueue {
         #[command(subcommand)]
-        command: queue::Command,
+        command: message_queue::Command,
     },
     /// Read queue items.
     Read {
@@ -37,7 +37,7 @@ pub enum Command {
     },
     /// Spawn an agent completion (open a streaming run as a child of this caller).
     Spawn(spawn::Command),
-    /// Client-side agent tags — lookup / add.
+    /// Client-side agent tags ÃƒÂ¢Ã¢â€šÂ¬Ã¢â‚¬Â lookup / add.
     Tags {
         #[command(subcommand)]
         command: tags::Command,
@@ -74,8 +74,8 @@ pub enum Request {
     PublishRequestSchema(publish::request_schema::Request),
     #[schemars(title = "PublishResponseSchema")]
     PublishResponseSchema(publish::response_schema::Request),
-    #[schemars(title = "Queue")]
-    Queue(queue::Request),
+    #[schemars(title = "MessageQueue")]
+    MessageQueue(message_queue::Request),
     #[schemars(title = "Read")]
     Read(read::Request),
     #[schemars(title = "Spawn")]
@@ -121,8 +121,8 @@ pub enum ResponseItem {
     PublishRequestSchema(publish::request_schema::Response),
     #[schemars(title = "PublishResponseSchema")]
     PublishResponseSchema(publish::response_schema::Response),
-    #[schemars(title = "Queue")]
-    Queue(queue::ResponseItem),
+    #[schemars(title = "MessageQueue")]
+    MessageQueue(message_queue::ResponseItem),
     #[schemars(title = "Read")]
     Read(read::ResponseItem),
     #[schemars(title = "Spawn")]
@@ -152,7 +152,7 @@ impl crate::cli::command::CommandResponse for ResponseItem {
             ResponseItem::Publish(v) => v.into_mcp(),
             ResponseItem::PublishRequestSchema(v) => v.into_mcp(),
             ResponseItem::PublishResponseSchema(v) => v.into_mcp(),
-            ResponseItem::Queue(v) => v.into_mcp(),
+            ResponseItem::MessageQueue(v) => v.into_mcp(),
             ResponseItem::Read(v) => v.into_mcp(),
             ResponseItem::Spawn(v) => v.into_mcp(),
             ResponseItem::SpawnRequestSchema(v) => v.into_mcp(),
@@ -196,8 +196,8 @@ impl TryFrom<Command> for Request {
                 Some(publish::Schema::ResponseSchema(args)) =>
                     Ok(Request::PublishResponseSchema(publish::response_schema::Request::try_from(args)?)),
             },
-            Command::Queue { command } =>
-                Ok(Request::Queue(queue::Request::try_from(command)?)),
+            Command::MessageQueue { command } =>
+                Ok(Request::MessageQueue(message_queue::Request::try_from(command)?)),
             Command::Read { command } =>
                 Ok(Request::Read(read::Request::try_from(command)?)),
             Command::Spawn(cmd) => match cmd.schema {
@@ -229,7 +229,7 @@ impl crate::cli::command::CommandRequest for Request {
             Request::Publish(inner) => inner.into_command(),
             Request::PublishRequestSchema(inner) => inner.into_command(),
             Request::PublishResponseSchema(inner) => inner.into_command(),
-            Request::Queue(inner) => inner.into_command(),
+            Request::MessageQueue(inner) => inner.into_command(),
             Request::Read(inner) => inner.into_command(),
             Request::Spawn(inner) => inner.into_command(),
             Request::SpawnRequestSchema(inner) => inner.into_command(),
@@ -338,9 +338,9 @@ pub async fn execute<E: crate::cli::command::CommandExecutor>(
                     ResponseItem::PublishResponseSchema(value),
                 )))
             }
-            Request::Queue(req) => {
-                let inner = queue::execute(executor, req, agent_arguments).await?;
-                Box::pin(inner.map(|r| r.map(ResponseItem::Queue)))
+            Request::MessageQueue(req) => {
+                let inner = message_queue::execute(executor, req, agent_arguments).await?;
+                Box::pin(inner.map(|r| r.map(ResponseItem::MessageQueue)))
             }
             Request::Read(req) => {
                 let inner = read::execute(executor, req, agent_arguments).await?;
@@ -457,8 +457,8 @@ pub async fn execute_jq<E: crate::cli::command::CommandExecutor>(
                 let value = publish::response_schema::execute_jq(executor, req, jq, agent_arguments).await?;
                 Box::pin(crate::cli::command::StreamOnce::new(Ok(value)))
             }
-            Request::Queue(req) => {
-                let inner = queue::execute_jq(executor, req, jq, agent_arguments).await?;
+            Request::MessageQueue(req) => {
+                let inner = message_queue::execute_jq(executor, req, jq, agent_arguments).await?;
                 Box::pin(inner)
             }
             Request::Read(req) => {
