--- objectiveai-sdk-rs/src/cli/command/agents/queue/list/request_schema/mod.rs
@@ -1,68 +1,0 @@
-use crate::cli::command::CommandRequest;
-
-#[derive(Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize, schemars::JsonSchema)]
-#[schemars(rename = "cli.command.agents.queue.list.request_schema.Request")]
-pub struct Request {
-    pub path_type: Path,
-    pub jq: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize, schemars::JsonSchema)]
-#[schemars(rename = "cli.command.agents.queue.list.request_schema.Path")]
-pub enum Path {
-    #[serde(rename = "agents/queue/list/request_schema")]
-    AgentsQueueListRequestSchema,
-}
-
-#[derive(clap::Args)]
-pub struct Args {
-    #[arg(long)]
-    pub jq: Option<String>,
-}
-
-impl CommandRequest for Request {
-    fn into_command(&self) -> Vec<String> {
-        let mut argv: Vec<String> = vec!["agents", "queue", "list", "request-schema"]
-            .into_iter()
-            .map(String::from)
-            .collect();
-        if let Some(jq) = &self.jq {
-            argv.push("--jq".to_string());
-            argv.push(jq.clone());
-        }
-        argv
-    }
-}
-
-pub type Response = crate::cli::command::ResponseSchema;
-
-impl TryFrom<Args> for Request {
-    type Error = crate::cli::command::FromArgsError;
-    fn try_from(args: Args) -> Result<Self, Self::Error> {
-        Ok(Self {
-            path_type: Path::AgentsQueueListRequestSchema,
-            jq: args.jq,
-        })
-    }
-}
-
-#[cfg(feature = "cli-executor")]
-pub async fn execute<E: crate::cli::command::CommandExecutor>(
-    executor: &E,
-    mut request: Request,
-    agent_arguments: Option<&crate::cli::command::AgentArguments>,
-) -> Result<Response, E::Error> {
-    request.jq = None;
-    executor.execute_one(request, agent_arguments).await
-}
-
-#[cfg(feature = "cli-executor")]
-pub async fn execute_jq<E: crate::cli::command::CommandExecutor>(
-    executor: &E,
-    mut request: Request,
-    jq: String,
-    agent_arguments: Option<&crate::cli::command::AgentArguments>,
-) -> Result<serde_json::Value, E::Error> {
-    request.jq = Some(jq);
-    executor.execute_one(request, agent_arguments).await
-}
